logger/requirement.txt
azure-monitor-ingestion
azure-identity